split index.d.ts
