+++ types/puppeteer/Coverage.d.ts
@@ -1,0 +1,25 @@
+export interface Coverage {
+  startCSSCoverage(options?: StartCoverageOptions): Promise<void>;
+  startJSCoverage(options?: StartCoverageOptions): Promise<void>;
+  stopCSSCoverage(): Promise<CoverageEntry[]>;
+  stopJSCoverage(): Promise<CoverageEntry[]>;
+}
+
+export interface CoverageEntry {
+  url: string;
+  text: string;
+  ranges: Array<{ start: number, end: number }>;
+}
+
+export interface StartCoverageOptions {
+  /**
+   * Whether to reset coverage on every navigation.
+   * @default true
+   */
+  resetOnNavigation?: boolean;
+  /**
+   * Whether anonymous scripts generated by the page should be reported.
+   * @default false
+   */
+  reportAnonymousScripts?: boolean;
+}
