+++ types/puppeteer/TaskQueue.d.ts
@@ -1,0 +1,3 @@
+export interface TaskQueue {
+    postTask(task: Function): Promise<any>;
+}
